qol improvements to reverse related fields
fixes to form file field

check isempty == isempty in forms.Haschanged

remove spammy log messages

add IsZero method to SimpleStoredObject

add users.group & users.permission reverse fields

change tests

add benchmarks

export NewRelatedQuerySet

+++ queries/test/benchmarks/one_to_one_test.go
@@ -50,8 +50,8 @@ func (t *BenchmarkO2OMain) FieldDefs(ctx context.Context) attrs.Definitions {
 
 type BenchmarkO2OThrough struct {
 	ID          uint64
-	SourceModel uint16
-	TargetModel uint16
+	SourceModel uint32
+	TargetModel uint32
 }
 
 func (t *BenchmarkO2OThrough) Fields() []attrs.Field {
